requirements.txt
# Core ML Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# NVIDIA NeMo (install from source for latest features)
# Install separately: pip install "nemo_toolkit[asr,tts] @ git+https://github.com/NVIDIA/NeMo.git"

# Lightning / PTL
pytorch-lightning>=2.2.0,<3

# Transformers and PEFT
transformers>=4.44.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Data Processing
datasets>=2.14.0
soundfile>=0.12.0
librosa>=0.10.0
pydub>=0.25.1
numpy>=1.24.0
scipy>=1.11.0
jiwer>=3.0.0,<4

# Training Infrastructure
wandb>=0.15.0
runpod>=1.5.0
python-dotenv>=1.0.0

# Utilities
omegaconf>=2.3.0
hydra-core>=1.3.0
tqdm>=4.66.0
pytest>=8.0.0
